backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pymongo>=4.0
supabase
celery
PyMuPDF
pytest
requests
mongomock
mock
redis
